requirements.txt
evaluate==0.3.0
flake8==5.0.4
isort==5.10.1
jiwer==2.5.1
pytorch-lightning==1.7.7
torchaudio==0.12.1
whisper @ git+https://github.com/openai/whisper.git@9e653bd0ea0f1e9493cb4939733e9de249493cfb
gdown==4.4.0
fastapi
uvicorn[standard]
python-multipart
loguru==0.7.0
# PyAudio==0.2.13
# SpeechRecognition==3.10.0
# sounddevice==0.4.6
# wikipedia==1.4.0
# pyttsx3==2.90
# dbus-python==1.3.2
